Core: Fix bulk loading with scala 2.12
When allLoader is not specified when building a LoadingCache or an
AsyncLoadingCache the method getAll is broken on scala 2.12 and
throw an java.lang.UnsupportedOperationException.

Indeed Caffeine use reflection to know if loadAll is overridden and
it seems that scala 2.12 bytecode generation break this, see
https://github.com/ben-manes/caffeine/blob/v2.3.3/caffeine/src/main/java/com/github/benmanes/caffeine/cache/LocalLoadingCache.java#L54.

To fix the issue this commit implements the adapters in java
instead of scala, thus when not providing an allLoader the loadAll
method is not seen overridden.